requirements.txt
numpy==1.22.4
pandas==1.3.4
mysql-connector-python==8.0.30
google-api-python-client==2.95.0
google-auth==2.22.0
google-analytics-data==0.17.0 
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-api-core==2.11.1